WIP: Refactor to use Models, Rename Project
Overhauling the entire project to add support for the majority of TMDB's REST API.

All entities are now managed via model classes and many resolvers for async fields are now static methods of these models. This is to help co-locate resolvers and REST response transformation, which was previously split up between resolver functions and the REST DataSource.

As a result of this refactor, there is now vastly improved typings throughout the code base.

Added support for a wide range of new data and arguments, notably cascading localization support. When a language is specified on a query or field, all nested descendants will be translated (where available) until a new language is set. This is designed to make localization of requests as simple as providing a single top-level query argument.

There are now two new top-level queries which provide powerful new ways to search for movies and tv shows, `discoverMovies` and `discoverTV`. Please take a look at the updated schema docs in the graphql playground to learn more.

All of this is currently a work in progress! Because of all the breaking changes, eventually this will be published as v3.0.0 under a new package name.