Drop support for Serverless Framework
It's been a pain in the ass to maintain, and since this is a hobby project I don't have the time nor energy to support a deployment target I no longer use.

This package exports everything you need to configure an Apollo server for whichever flavor of Apollo Server you like, so if you want to figure out how to deploy it directly to AWS you can. Personally, I'd recommend going with Netlify or Vercel functions as they take care a lot of the deployment headaches for you.